requirements.txt
python-dotenv
pydantic
requests
websockets
psycopg2-binary
rank-bm25
rapidfuzz
